requirements.txt
click
pyobjc-framework-Quartz
Pillow